Refactor: Single API file into separate using claude ai
