+++ frontend/src/components/auth/SignInForm.tsx
@@ -4,7 +4,7 @@ import { ChevronLeftIcon, EyeCloseIcon, EyeIcon } from "../../icons";
 import Label from "../form/Label";
 import Input from "../form/input/InputField";
 import Button from "../ui/button/Button";
-import { fetchLogin } from "../../API/API";
+import { fetchLogin } from "../../API";
 
 export default function SignInForm() {
   const [showPassword, setShowPassword] = useState(false);
